requirements.txt
asgiref
certifi
charset-normalizer==2.1.1
cloudinary
Django
django-admin-interface
django-cloudinary-storage
django-colorfield
django-cors-headers
django-flat-responsive
django-flat-theme
django-jazzmin
djangorestframework
djongo==1.3.6
dnspython==2.2.1
idna
Pillow
pymongo==3.12.3
python-magic
pytz
requests
six
sqlparse==0.2.4
tzdata
urllib3==1.26.5
PyJWT
